Finished version. ENV var setup missing
