--- mysite/classlink/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path("", views.index, name="index"),
-]
