--- mysite/classlink/views.py
@@ -1,7 +1,0 @@
-from django.shortcuts import render
-
-from django.http import HttpResponse
-
-
-def index(request):
-    return HttpResponse("Hello, world. You're at the classlink index.")
