+++ mysite/home/templates/home/student_attendance.html
