--- mysite/snappy/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SnappyConfig(AppConfig):
-    name = 'snappy'
