--- mysite/snappy/context_processors.py
@@ -1,7 +1,0 @@
-from django.conf import settings as django_settings
-
-
-def settings(request):
-    return {
-        'settings': django_settings,
-    }
