--- mysite/snappy/static/favicon.notes
@@ -1,4 +1,0 @@
-Favicon made using
-
-https://favicon.io/favicon-converter/
-
