--- mysite/snappy/static/home/styles.css
@@ -1,68 +1,0 @@
-/* custom-styles.css */
-
-
-.jumbotron {
-  background-color: #f8f9fa;
-  text-align: center;
-  border: none;
-  border-radius: .3rem;
-}
-
-
-.text-primary {
-  color: #007bff;
-}
-
-.text-secondary {
-  color: #808080;
-}
-
-.card {
-  border: 1px solid #007bff;
-}
-
-.card-body {
-  text-align: center;
-}
-
-/* Icons within card body
-.card-body i {
-  color: #007bff;
-  margin-bottom: .75rem;
-}
-
-/* Card title styling
-.card-title {
-  color: #007bff;
-}
-
-
-.card-text {
-  color: #495057;
-}
-
-/* Buttons */
-.btn-primary {
-  background-color: #007bff;
-  border-color: #007bff;
-  color: #ffffff;
-}
-
-.btn-outline-primary {
-  border-color: #007bff;
-  color: #007bff;
-}
-
-.btn-outline-primary:hover {
-  background-color: #007bff;
-  color: #ffffff;
-}
-
-
-.container.mt-5 {
-  margin-top: 3rem !important;
-}
-
-
-@media (max-width: 768px) {
-}
