--- mysite/snappy/templates/registration/login.html
@@ -1,41 +1,0 @@
-<html>
-  {% block content %}
- <head>
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css">
-  </head>
-    <title>Log In</title>
-  <div class="container mt-5 center">
-      <div class="row justify-content-center">
-          <div class="col-md-6">
-              <div class="card border-primary">
-                  <div class="card-body">
-                      <h3 class="">Login</h3>
-                      {% if form.errors %}
-                          <div class="alert alert-danger" role="alert">
-                              Your username and password didn't match. Please try again.
-                          </div>
-                      {% endif %}
-                      <form method="post" action="{% url 'login' %}">
-                          {% csrf_token %}
-                          <div class="form-group">
-                              {{ form.username.label_tag }}
-                              {{ form.username }}
-                              {{ form.username.errors }}
-                          </div>
-                          <div class="form-group">
-                              {{ form.password.label_tag }}
-                              {{ form.password }}
-                              {{ form.password.errors }}
-                          </div>
-                          <input type="submit" class="btn btn-outline-primary btn-block" value="Login" />
-                          <input type="hidden" name="next" value="{{ next }}" />
-                      </form>
-                  </div>
-              </div>
-          </div>
-      </div>
-  </div>
-  
-  {% endblock %}
-</html>
