--- mysite/snappy/templates/snappy/attendance.html
@@ -1,66 +1,0 @@
-{% block content %}
-
-<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css">
-
-<input type="hidden" id="classId" value="{{ selected_class.id }}">
-<input type="hidden" id="classId" value="{{ current_session.session_number }}">
-
-
-<div class="container mt-5">
-    <div class="jumbotron bg-light text-center">
-        <h1 class="display-4 text-primary">{{ selected_class }}</h1>
-        <p class="lead">Manage attendance for your class.</p>
-        <h1 class="display-4 text-secondary">Attendance Code: <span id="attendanceCode"></span> </h1>
-        
-        <a href="{% url 'attendance_management' %}" class="btn btn-secondary">Check Attendance</a>
-        <hr class="my-4">
-    </div>
-</div>
-{% endblock %}
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-
-
-<script>
-    $(document).ready(function() {
-        // Function to delete old codes from Azure Table Storage
-        function deleteOldCodes() {
-            $.ajax({
-                url: 'https://tempcodes.azurewebsites.net/api/DeleteCodesFunction?code=bjDNB3x9NRcJtChu5YvEuSOT__c3jdmRnKhDmoIDwjX2AzFuXYyQbw==',
-                type: 'GET',
-                success: function(response, statusText, xhr) {
-                    console.log('Old codes deleted, status:', xhr.status);
-                },
-                error: function(error) {
-                    console.log('Error deleting old codes:', error);
-                }
-            });
-        }
-
-        // Function to update the attendance code from Azure Function
-        function updateAttendanceCode() {
-            deleteOldCodes(); // Delete old codes before updating the new code
-
-            $.ajax({
-                url: 'https://tempcodes.azurewebsites.net/api/GenerateCodeFunction?code=74pbScwXMtLf6JDxQy6s8MpIEHTHXke6q-7mP5BX-JkIAzFumSyvkQ==',
-                type: 'GET',
-                success: function(response) {
-                    console.log('New code fetched:', response.code);
-                    $('#attendanceCode').text(response.code);
-                },
-                error: function(error) {
-                    console.log('Error fetching new code:', error);
-                }
-            });
-        }
-
-        // Schedule the functions
-        setInterval(updateAttendanceCode, 30000); // Update code every 30 seconds
-
-        // Initial calls
-        updateAttendanceCode();
-    });
-</script>
