--- mysite/snappy/templates/snappy/attendance_management.html
@@ -1,57 +1,0 @@
-{% block content %}
-
-<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css">
-
-<div class="container mt-5">
-    <div class="jumbotron bg-light text-center">
-        <h1 class="display-4 text-primary">{{ course.course_name }}</h1>
-        <p class="lead">Manage attendance for your class. Session Number: {{ current_session.session_number }}</p>
-        <hr class="my-4">
-    </div>
-    <form method="post" action="{% url 'attendance_management' %}"> 
-        {% csrf_token %}
-
-        <div class="row justify-content-center">
-            <div class="col-md-8">
-                <table class="table mt-4">
-                    <thead class="thead-light">
-                        <tr>
-                            <th>Student Name</th>
-                            <th>Status</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {% for record in attendance_records %}
-                        <tr>
-                            <td>{{ record.name }}</td> 
-                            <td>
-                                <select class="form-control status-select" name="status_{{ record.id }}">
-                                    <option value="present" {% if record.status == 'present' %}selected{% endif %}>Present</option>
-                                    <option value="absent" {% if record.status == 'absent' %}selected{% endif %}>Absent</option>
-                                    <option value="late" {% if record.status == 'late' %}selected{% endif %}>Late</option>
-                                </select>
-                            </td>
-                        </tr>
-                        {% empty %}
-                        <tr>
-                            <td colspan="2">No students enrolled in this class.</td>
-                        </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        </div>
-        <div class="row justify-content-center">
-            <div class="col-md-8 text-center">
-                <button type="submit" class="btn btn-primary">Update Attendance</button>
-            </div>
-        </div>
-    </form>
-</div>
-
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-
-{% endblock %}
