--- mysite/snappy/templates/snappy/confirmation.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-{% block content %}
-<html>
-    <head>
-            <title>Submission Success</title>
-            <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-            <style>
-                .container {{
-                    margin-top: 50px;
-                }}
-            </style>
-        </head>
-        <body>
-            <div class="container">
-                <div class="alert alert-success" role="alert">
-                    <h4 class="alert-heading">Successfully Submitted!</h4>
-                    <p>Class: {{class}}</p>
-                    <hr>
-                    {%if is_professor %}
-                    <p class="mb-0">Class attendance was submitted correctly.</p>
-                    {% else %}
-                    <p class="mb-0">Your attendance code has been submitted successfully.</p>
-                    {% endif %}
-                </div>
-                <a href="{% url 'home' %}" class="btn btn-primary">Home</a>
-            </div>
-        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-        <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-    </body>
-</html>
-{% endblock %}
