--- mysite/snappy/templates/snappy/professor_dashboard.html
@@ -1,40 +1,0 @@
-{% load static %}
-
-{% block content %}
-
-<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css">
-<link rel="stylesheet" href="{% static 'styles.css' %}">
-
-<div class="container mt-5">
-    <div class="jumbotron bg-light text-center">
-        <h1 class="display-4 text-primary">Professor <span class ="display-4 text-secondary" > {{professor.name}}</span></h1>
-        <p class="lead">Select a class to manage attendance codes.</p>
-        <hr class="my-4">
-    </div>
-
-    <div class="row justify-content-center">
-        <div class="col-md-6">
-            <form action="{% url 'professor_dashboard' %}" method="post">
-                {% csrf_token %}
-                <div class="form-group">
-                    <label for="selectClass">Select Class:</label>
-                    <select class="form-control" id="selectClass" name="selected_class">
-                        {% for course in classes %}
-                            <option value="{{ course.id }}">{{ course.course_name }}</option>
-                        {% empty %}
-                            <option>No courses available</option>
-                        {% endfor %}
-                    </select>
-                </div>
-                <button type="submit" class="btn btn-primary">Display Codes</button>
-            </form>
-        </div>
-    </div>
-</div>
-
-{% endblock %}
-
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
