--- mysite/snappy/templatetags/app_tags.py
@@ -1,21 +1,0 @@
-from hashlib import md5
-from django import template
-
-# https://brobin.me/blog/2016/07/super-simple-django-gravatar/
-
-# A "gravatar" is a globally recognized avatar that is based on email address
-# People must register their email address and then upload a gravatar
-# If an email address has no gravatar, a generic image is put in its place
-
-# To use the gravatar filter in a template include
-# {% load app_tags %}
-
-register = template.Library()
-
-
-@register.filter(name='gravatar')
-def gravatar(user, size=35):
-    email = str(user.email.strip().lower()).encode('utf-8')
-    email_hash = md5(email).hexdigest()
-    url = "//www.gravatar.com/avatar/{0}?s={1}&d=identicon&r=PG"
-    return url.format(email_hash, size)
